fix: multibyte instructions use variable-length integer encoding

+++ src/core/utils.rs
@@ -10,7 +10,7 @@ pub fn print_beautiful_instruction_name_1_byte(first_byte: u8, pc: usize) {
 }
 
 #[cfg(debug_assertions)]
-pub fn print_beautiful_fc_extension(second_byte: u8, pc: usize) {
+pub fn print_beautiful_fc_extension(second_byte: u32, pc: usize) {
     use crate::core::reader::types::opcode::fc_extension_opcode_second_byte_to_str;
 
     trace!(
